+++ stripe/src/test/java/com/stripe/android/model/PaymentIntentTest.java
@@ -16,7 +16,9 @@ import java.util.Map;
 import static com.stripe.android.testharness.JsonTestUtils.assertJsonEqualsExcludingNulls;
 import static com.stripe.android.testharness.JsonTestUtils.assertMapEquals;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 @RunWith(RobolectricTestRunner.class)
 public class PaymentIntentTest {
@@ -36,7 +38,7 @@ public class PaymentIntentTest {
             "  \"currency\": \"usd\",\n" +
             "  \"description\": \"Example PaymentIntent charge\",\n" +
             "  \"livemode\": false,\n" +
-            "  \"next_source_action\": null,\n" +
+            "  \"next_action\": null,\n" +
             "  \"receipt_email\": null,\n" +
             "  \"shipping\": null,\n" +
             "  \"source\": \"src_1CkiC3LENEVhOs7YMSa4yx4G\",\n" +
@@ -57,15 +59,17 @@ public class PaymentIntentTest {
             "  \"currency\": \"usd\",\n" +
             "  \"description\": \"Example PaymentIntent charge\",\n" +
             "  \"livemode\": false,\n" +
-            "  \"next_source_action\": {" +
+            "  \"next_action\": {" +
             "       \"type\": \"authorize_with_url\"," +
-            "           authorize_with_url: {" +
-            "           url: \""+ BAD_URL +"\" } " +
+            "           \"authorize_with_url\": {" +
+            "             \"url\": \""+ BAD_URL +"\"," +
+            "             \"return_url\": \"yourapp://post-authentication-return-url\"" +
+            "           } " +
             "       },\n" +
             "  \"receipt_email\": null,\n" +
             "  \"shipping\": null,\n" +
             "  \"source\": \"src_1CkiC3LENEVhOs7YMSa4yx4G\",\n" +
-            "  \"status\": \"requires_source_action\"\n" +
+            "  \"status\": \"requires_action\"\n" +
             "}\n";
 
     private static final String PAYMENT_INTENT_WITH_PAYMENT_METHODS_JSON = "{\n" +
@@ -133,7 +137,8 @@ public class PaymentIntentTest {
             "\t\"next_action\": {\n" +
             "\t\t\"type\": \"redirect_to_url\",\n" +
             "\t\t\"redirect_to_url\": {\n" +
-            "\t\t\t\"url\": \"https://example.com/redirect\"\n" +
+            "\t\t\t\"url\": \"https://example.com/redirect\",\n" +
+            "\t\t\t\"return_url\": \"yourapp://post-authentication-return-url\"\n" +
             "\t\t}\n" +
             "\t}\n" +
             "}";
@@ -145,7 +150,8 @@ public class PaymentIntentTest {
             "\t\"next_action\": {\n" +
             "\t\t\"type\": \"authorize_with_url\",\n" +
             "\t\t\"authorize_with_url\": {\n" +
-            "\t\t\t\"url\": \"https://example.com/redirect\"\n" +
+            "\t\t\t\"url\": \"https://example.com/redirect\",\n" +
+            "\t\t\t\"return_url\": \"yourapp://post-authentication-return-url\"\n" +
             "\t\t}\n" +
             "\t}\n" +
             "}";
@@ -197,9 +203,8 @@ public class PaymentIntentTest {
                 PAYMENT_INTENT_WITH_SOURCE_WITH_BAD_AUTH_URL_JSON);
         assertNotNull(paymentIntent);
 
-        final Uri authUrl = paymentIntent.getAuthorizationUrl();
+        final Uri authUrl = paymentIntent.getRedirectUrl();
         assertNotNull(authUrl);
-
         assertEquals(BAD_URL, authUrl.getEncodedPath());
     }
 
@@ -208,6 +213,7 @@ public class PaymentIntentTest {
         final PaymentIntent paymentIntent = PaymentIntent
                 .fromString(PARTIAL_PAYMENT_INTENT_WITH_REDIRECT_URL_JSON);
         assertNotNull(paymentIntent);
+        assertTrue(paymentIntent.requiresAction());
         final Uri redirectUrl = paymentIntent.getRedirectUrl();
         assertNotNull(redirectUrl);
         assertEquals("https://example.com/redirect", redirectUrl.toString());
@@ -235,6 +241,7 @@ public class PaymentIntentTest {
         final PaymentIntent paymentIntent =
                 PaymentIntent.fromString(PAYMENT_INTENT_WITH_PAYMENT_METHODS_JSON);
         assertNotNull(paymentIntent);
+        assertFalse(paymentIntent.requiresAction());
         assertEquals("card", paymentIntent.getPaymentMethodTypes().get(0));
     }
 }